Reorganize native syscall C code under sys hierarchy
