+++ src/AiVM.Core/native/include/aivm_runtime.h
@@ -2,7 +2,7 @@
 #define AIVM_RUNTIME_H
 
 #include "aivm_program.h"
-#include "aivm_syscall.h"
+#include "sys/aivm_syscall.h"
 #include "aivm_vm.h"
 
 int aivm_execute_program(const AivmProgram* program, AivmVm* vm_out);
